AGENT_requirements.txt
torch==2.6.0+cu128 --index-url https://download.pytorch.org/whl/cu128
torchaudio==2.6.0+cu128
torchvision==0.21.0+cu128
transformers
wandb
logging
bitsandbytes
accelerate>=0.26.0
sentence_transformers